refactor: rename styles.module for Banner, Reward & Stake
