+++ src/pages/PageStake/PageStake.module.scss
