+++ src/pages/PageStake/components/Banner/Banner.module.scss
