+++ src/pages/PageStake/components/Reward/Reward.module.scss
