+++ src/pages/PageStake/components/Stake/Stake.module.scss
